requirements.txt
streamlit>=1.36.0
langchain>=0.2.10
langchain-community>=0.2.10
openai>=1.40.0
chromadb==0.4.24
pysqlite3-binary>=0.5.2
faiss-cpu>=1.8.0
pypdf>=4.2.0
sentence-transformers>=3.0.1
python-dotenv>=1.0.1
tiktoken>=0.7.0
numpy==1.26.4
google-generativeai>=0.6.0
langchain-google-genai>=1.0.7
transformers>=4.42.0

